backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.7.0
pydantic[email]
SQLAlchemy==2.0.30
httpx==0.27.0
python-multipart==0.0.9
